Close loop on turret finished and tested.

+++ STM32F405RGT6/User/app_chassis.c
@@ -7,10 +7,10 @@ extern int16_t measured_202_speed;
 extern int16_t measured_203_speed;
 extern int16_t measured_204_speed;
 
-extern int16_t measured_201_angle;
-extern int16_t measured_202_angle;
-extern int16_t measured_203_angle;
-extern int16_t measured_204_angle;
+extern float measured_201_angle;
+extern float measured_202_angle;
+extern float measured_203_angle;
+extern float measured_204_angle;
 
 const float v_p_201 = 1.0;
 const float v_i_201 = 0;
